# Changes to optimize gr methods