--- a/setup.cfg
+++ b/setup.cfg
@@ -50,7 +50,7 @@ python_requires = >=3.8
 install_requires =
     importlib-metadata; python_version<"3.8"
     biocframe>=0.5.11
-    iranges[optional]>=0.2.9
+    iranges[optional]>=0.2.10
     biocutils>=0.1.3
     numpy