Tags: #98 

Current tests: 

Files downloaded from http://catlas.org/wholemousebrain/#!/downloads:

```python
import biobear as bb
import polars as pl
from genomicranges import GenomicRanges

bed_file = "mousebrain.union.cCRE.bed"
session = bb.new_session()
bed = session.read_bed_file(bed_file, bb.BEDReadOptions(n_fields=3))
bed_df = bed.to_polars()
bed_df.shape
## (491818, 3)

bw_file = "317_Astro-CB_NN.bw"
bigwig = session.read_bigwig_file(bw_file)
bigwig_df = bigwig.to_polars()
bigwig_df.shape
## (14724918, 4)
```

bed_gr:
```python
%%time

bed_gr = GenomicRanges.from_polars(bed_df.rename({"reference_sequence_name": "seqnames", "start": "starts", "end": "ends"}))
CPU times: user 61.4 ms, sys: 6.98 ms, total: 68.4 ms
Wall time: 68.4 ms
```

bigwig_gr:

```python
%%time
# %%prun
updated_bw = bigwig_df.rename({"name": "seqnames", "start": "starts", "end": "ends"})
bigwig_gr = GenomicRanges.from_polars(updated_bw)
CPU times: user 1.91 s, sys: 692 ms, total: 2.6 s
Wall time: 3.09 s
```

Finally perform intersect:

```python
%%time

bed_bigwig_gr = bed_gr.intersect(bigwig_gr)
CPU times: user 1min 54s, sys: 1.48 s, total: 1min 56s
Wall time: 1min 57s
```